requirements.txt
fastapi
uvicorn
pydantic
requests
pytest
factory-boy
behave